+++ Package.swift
@@ -17,7 +17,7 @@ let package = Package(
         ),
     ],
     dependencies: [
-        .package(url: "https://github.com/gifton/VectorCore", from: "0.1.6")
+        .package(url: "https://github.com/gifton/VectorCore", from: "0.2.0")
     ],
     targets: [
         .target(